Update scatterplot.Rmd (#198)
* Update scatterplot.Rmd

I understand that there is some overlap here with the "when to use" section. I think it could be beneficial to be consistent and either have a when to use section in every "chapter" or let it be subsumed into theory. Maybe this is too knitpicky but I think there is more to be said and it would be good to have consistency.

* Update scatterplot.Rmd

Co-Authored-By: leahparreztnik <44757731+leahparreztnik@users.noreply.github.com>